Split the run side of lib.rs into lib_run.rs
